// client/archival-material-client/src/layout/RoutesEnum.ts
enum RoutesEnum {
    LOGIN = "/login",
    REGISTER = "/register",
    HOME = "/",
    FORGOT_PASSWORD = "/forgot-password",
    CHANGE_PASSWORD = "/change-password",
    RESET_PASSWORD = "/reset-password",
    ARCHIVAL_MATERIAL = "/archival-materials",
    ARCHIVAL_MATERIAL_PREVIEW = "/archival-materials/preview",
    PROFILE = "/profile",
    UNAUTHORIZED = "/unauthorized",
    SERVER_DOWN = "/server-down",
    NOT_FOUND = "/not-found",
    CREDIT = "/credit",
    OTHER_PROJECTS = "/other-projects"
}

export default RoutesEnum;